UHF-2506: Set TPR Service's errand_services field translatable
The API contains services and errand services in different languages.
However, not all content is translated. A service may refer to a
certain errand service, for example, in the Finnish version, but the
errand service might not be translated to other languages and,
consequently, the service might not refer to the same errand service in
other languages.

Without translating the errand_services field, the value from the
lastly imported language version overrides the previously imported
values.