UHF-2834: Always use Finnish as service's default language
There was a problem when importing services.

Assuming there was an existing service entity in Finnish with
translations in Swedish and English. Then, if fetching the Finnish
source data, for some reason, failed, the English version became the
default one. This caused the Finnish version to disappear from the UI,
as it was no longer included with the latest revision.

At this point, the English version had the entity references which were
previously added to the Finnish version.

After this, when re-importing the service without failing to get the
data, the Finnish version became once again the default language version.
Now there were all three language versions, but the English translation
had lost all its previously added entity references.