Fix failing UnitListTest
Although the units are initially imported using the fixture file,
updating the selected units from the admin UI seems to use the actual
units API. Apparently the API was changed so that there is no longer
a matching unit ID when trying to update the imported unit. Because of
this, the unit was not updated and the test was failing.